同步

+++ lib/Maintenance/Tasks/ScheduledTasksTask.php
@@ -19,6 +19,7 @@ use Pimcore\Model\Asset;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\Document;
 use Pimcore\Model\Schedule\Task\Listing;
+use Pimcore\Model\User;
 use Pimcore\Model\Version;
 use Psr\Log\LoggerInterface;
 
@@ -47,13 +48,14 @@ final class ScheduledTasksTask implements TaskInterface
         $tasks = $list->load();
 
         foreach ($tasks as $task) {
+            $taskUser = User::getById($task->getUserId());
+
             try {
                 if ($task->getCtype() === 'document') {
                     $document = Document::getById($task->getCid());
                     if ($document instanceof Document) {
-                        if ($task->getAction() === 'publish-version' && $task->getVersion()) {
-                            try {
-                                $version = Version::getById($task->getVersion());
+                        if ($task->getAction() === 'publish-version' && $task->getVersion() && $document->isAllowed('publish', $taskUser) && $document->isAllowed('versions', $taskUser)) {
+                            if ($version = Version::getById($task->getVersion())) {
                                 $document = $version->getData();
                                 if ($document instanceof Document) {
                                     $document->setPublished(true);
@@ -61,16 +63,16 @@ final class ScheduledTasksTask implements TaskInterface
                                 } else {
                                     $this->logger->error('Schedule\\Task\\Executor: Could not restore document from version data.');
                                 }
-                            } catch (\Exception $e) {
+                            } else {
                                 $this->logger->error('Schedule\\Task\\Executor: Version [ '.$task->getVersion().' ] does not exist.');
                             }
-                        } elseif ($task->getAction() === 'publish') {
+                        } elseif ($task->getAction() === 'publish' && $document->isAllowed('publish', $taskUser)) {
                             $document->setPublished(true);
                             $document->save();
-                        } elseif ($task->getAction() === 'unpublish') {
+                        } elseif ($task->getAction() === 'unpublish' && $document->isAllowed('unpublish', $taskUser)) {
                             $document->setPublished(false);
                             $document->save();
-                        } elseif ($task->getAction() === 'delete') {
+                        } elseif ($task->getAction() === 'delete' && $document->isAllowed('delete', $taskUser)) {
                             $document->delete();
                         }
                     }
@@ -78,19 +80,18 @@ final class ScheduledTasksTask implements TaskInterface
                     $asset = Asset::getById($task->getCid());
 
                     if ($asset instanceof Asset) {
-                        if ($task->getAction() === 'publish-version' && $task->getVersion()) {
-                            try {
-                                $version = Version::getById($task->getVersion());
+                        if ($task->getAction() === 'publish-version' && $task->getVersion() && $asset->isAllowed('publish', $taskUser) && $asset->isAllowed('versions', $taskUser)) {
+                            if ($version = Version::getById($task->getVersion())) {
                                 $asset = $version->getData();
                                 if ($asset instanceof Asset) {
                                     $asset->save();
                                 } else {
                                     $this->logger->error('Schedule\\Task\\Executor: Could not restore asset from version data.');
                                 }
-                            } catch (\Exception $e) {
+                            } else {
                                 $this->logger->error('Schedule\\Task\\Executor: Version [ '.$task->getVersion().' ] does not exist.');
                             }
-                        } elseif ($task->getAction() === 'delete') {
+                        } elseif ($task->getAction() === 'delete' && $asset->isAllowed('delete', $taskUser)) {
                             $asset->delete();
                         }
                     }
@@ -98,9 +99,8 @@ final class ScheduledTasksTask implements TaskInterface
                     $object = DataObject::getById($task->getCid());
 
                     if ($object instanceof DataObject) {
-                        if ($task->getAction() === 'publish-version' && $task->getVersion()) {
-                            try {
-                                $version = Version::getById($task->getVersion());
+                        if ($task->getAction() === 'publish-version' && $task->getVersion() && $object->isAllowed('publish', $taskUser) && $object->isAllowed('versions', $taskUser)) {
+                            if ($version = Version::getById($task->getVersion())) {
                                 $object = $version->getData();
                                 if ($object instanceof DataObject\AbstractObject) {
                                     $object->setPublished(true);
@@ -108,16 +108,16 @@ final class ScheduledTasksTask implements TaskInterface
                                 } else {
                                     $this->logger->error('Schedule\\Task\\Executor: Could not restore object from version data.');
                                 }
-                            } catch (\Exception $e) {
+                            } else {
                                 $this->logger->error('Schedule\\Task\\Executor: Version [ '.$task->getVersion().' ] does not exist.');
                             }
-                        } elseif ($task->getAction() === 'publish') {
+                        } elseif ($task->getAction() === 'publish' && $object->isAllowed('publish', $taskUser)) {
                             $object->setPublished(true);
                             $object->save();
-                        } elseif ($task->getAction() === 'unpublish') {
+                        } elseif ($task->getAction() === 'unpublish' && $object->isAllowed('unpublish', $taskUser)) {
                             $object->setPublished(false);
                             $object->save();
-                        } elseif ($task->getAction() === 'delete') {
+                        } elseif ($task->getAction() === 'delete' && $object->isAllowed('delete', $taskUser)) {
                             $object->delete();
                         }
                     }
