+++ models/DataObject/Objectbrick/Dao.php
@@ -20,8 +20,8 @@ namespace Pimcore\Model\DataObject\Objectbrick;
 use Pimcore\Model;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\DataObject\ClassDefinition\Data\CustomResourcePersistingInterface;
+use Pimcore\Model\DataObject\ClassDefinition\Data\LazyLoadingSupportInterface;
 use Pimcore\Model\DataObject\ClassDefinition\Data\ResourcePersistenceAwareInterface;
-use Pimcore\Tool;
 
 /**
  * @property \Pimcore\Model\DataObject\Objectbrick $model
@@ -36,14 +36,13 @@ class Dao extends Model\DataObject\Fieldcollection\Dao
      */
     public function load(DataObject\Concrete $object, $params = [])
     {
+        /** @var DataObject\ClassDefinition\Data\Objectbricks $fieldDef */
         $fieldDef = $object->getClass()->getFieldDefinition($this->model->getFieldname());
 
         $values = [];
 
         foreach ($fieldDef->getAllowedTypes() as $type) {
-            try {
-                $definition = DataObject\Objectbrick\Definition::getByKey($type);
-            } catch (\Exception $e) {
+            if (!$definition = DataObject\Objectbrick\Definition::getByKey($type)) {
                 continue;
             }
 
@@ -64,14 +63,10 @@ class Dao extends Model\DataObject\Fieldcollection\Dao
                 $brick->setObject($object);
 
                 foreach ($fieldDefinitions as $key => $fd) {
-                    if ($fd instanceof CustomResourcePersistingInterface || method_exists($fd, 'load')) {
-                        if (!$fd instanceof CustomResourcePersistingInterface) {
-                            Tool::triggerMissingInterfaceDeprecation(get_class($fd), 'load', CustomResourcePersistingInterface::class);
-                        }
-
+                    if ($fd instanceof CustomResourcePersistingInterface) {
                         $doLoad = true;
 
-                        if ($fd instanceof  DataObject\ClassDefinition\Data\Relations\AbstractRelations) {
+                        if ($fd instanceof LazyLoadingSupportInterface) {
                             if (!DataObject\Concrete::isLazyLoadingDisabled() && $fd->getLazyLoading()) {
                                 $doLoad = false;
                             }
@@ -93,10 +88,7 @@ class Dao extends Model\DataObject\Fieldcollection\Dao
                             }
                         }
                     }
-                    if ($fd instanceof ResourcePersistenceAwareInterface || method_exists($fd, 'getDataFromResource')) {
-                        if (!$fd instanceof ResourcePersistenceAwareInterface) {
-                            Tool::triggerMissingInterfaceDeprecation(get_class($fd), 'getDataFromResource', ResourcePersistenceAwareInterface::class);
-                        }
+                    if ($fd instanceof ResourcePersistenceAwareInterface) {
                         if (is_array($fd->getColumnType())) {
                             $multidata = [];
                             foreach ($fd->getColumnType() as $fkey => $fvalue) {
@@ -118,7 +110,7 @@ class Dao extends Model\DataObject\Fieldcollection\Dao
                 $setter = 'set' . ucfirst($type);
 
                 if ($brick instanceof DataObject\DirtyIndicatorInterface) {
-                    $brick->markFieldDirty($key, false);
+                    $brick->markFieldDirty('_self', false);
                 }
 
                 $this->model->$setter($brick);
@@ -132,23 +124,19 @@ class Dao extends Model\DataObject\Fieldcollection\Dao
 
     /**
      * @param DataObject\Concrete $object
-     * @param $saveMode true if called from save method
-     *
-     * @return whether an insert should be done or not
+     * @param bool $saveMode true if called from save method
      */
     public function delete(DataObject\Concrete $object, $saveMode = false)
     {
         // this is to clean up also the inherited values
+
+        /** @var DataObject\ClassDefinition\Data\Objectbricks $fieldDef */
         $fieldDef = $object->getClass()->getFieldDefinition($this->model->getFieldname());
         foreach ($fieldDef->getAllowedTypes() as $type) {
-            try {
-                $definition = DataObject\Objectbrick\Definition::getByKey($type);
-            } catch (\Exception $e) {
-                continue;
+            if ($definition = DataObject\Objectbrick\Definition::getByKey($type)) {
+                $tableName = $definition->getTableName($object->getClass(), true);
+                $this->db->delete($tableName, ['o_id' => $object->getId()]);
             }
-
-            $tableName = $definition->getTableName($object->getClass(), true);
-            $this->db->delete($tableName, ['o_id' => $object->getId()]);
         }
     }
 }
